move pricing components into /about
