+++ src/presenters/pages/about/pricing.styl
@@ -1,4 +1,4 @@
-@import '../../components/constants.styl';
+@import '../../../components/constants.styl';
 
 .pricingBlocks
     display: grid
